requirements.txt
altair<5
google-cloud-storage==2.12.0
faiss-cpu==1.7.4
python-docx==1.0.1
langchain==0.0.325
streamlit-authenticator==0.2.3
streamlit==1.4.0
tika==2.6.0
click<=8.0.4
pydub==0.25.1
tiktoken==0.5.1
openai==0.28.1
yt-dlp==2023.10.13
protobuf==3.20